moves static data files to new folder
