+++ governance-aepp/src/utils/aeternity.js
@@ -2,7 +2,7 @@ import Aepp from '@aeternity/aepp-sdk/es/ae/aepp';
 import Util from './util';
 import registryContractSource from '../assets/contracts/RegistryInterface.aes';
 import {Universal} from "@aeternity/aepp-sdk/es/ae/universal";
-import settings from './settings';
+import settings from '../data/settings';
 
 const aeternity = {
   client: null,
